+++ crates/cli/src/commands/run.rs
@@ -335,6 +335,7 @@ pub async fn run(session: ProtoSession, args: RunArgs) -> AppResult {
                         backend: spec.backend,
                         req: resolved_version.to_unresolved_spec(),
                         res: Some(resolved_version.clone()),
+                        resolve_from_manifest: false,
                         read_lockfile: false,
                         write_lockfile: false,
                     }),